// src/common/enum/api-error-code.enum.ts
export enum ApiErrorCode {
  TIMEOUT = -1, // 系统繁忙
  SUCCESS = 0, // 成功
  USER_ID_INVALID = 10001, // 用户id无效
  ACCOUNT_INVALID = 10002,
  PASSWORD_INVALID = 10003,
  EMAIL_EXIST = 10004,
  ACCOUNT_DELETED = 10005,
  LOGIN_ERROR = 10010,
  NO_PERMISSION = 10011,
  INPUT_ERROR = 10012,
  NO_EXIST = 10013,
  PHONE_EXIST = 10014,
  INTERNAL_ERROR = 10015,
}